chore(various): Testing ATS setup
Changing the ATS setup to always upload to codecov.
Changing the ATS setup to have the multiple jobs like backend.
Separating CLI config from `codecov.yml` because otherwise it's invalid

Some very small - but real - changes that might go into sentry repo in the future.
This is just a test for the ATS setup.
This commit should run a mountain of tests, so fingers crossed that everything goes well.