ci: Re-enable codecov-ats
This reverts commit f002da89638ba2e8545638a8ec11c99f2ca179a4.
(reactivates codecov-ats workflow)

This workflow was removed by commit f002da89638ba2e8545638a8ec11c99f2ca179a4
because the `concurrency:` key was not idented properly.
This caused errors on `pre-commit run check-github-workflows --all-files`

Running that command also gave errors for the `strategy` portion, because apparently `matrix:` is required.
My understanding is that this is used for concurrency (having multiple paralel jobs, test sharding, for example),
but this is not the case here, so I just ripped it off completely.